api_gateway/requirements.txt
fastapi==0.70.0
httpx==0.21.1
python-jose==3.3.0
uvicorn==0.15.0
